add barebones extension
this can add about 1b numbers together in 30s. it executes most of the workload on a single core and then spends the last 5s in multicore.

this sorta shows that duckdb is quite fast. i think performance decreases with the number of columns and complexity of the columns but not sure exactly by how much